break up migration tests
