+++ tests/unit/test_migration/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2023 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Unit tests for migartion module."""
